get rid of attribute source optimizing visitor
